+++ spec/todoapp/app/models/employee.rb
@@ -1,0 +1,3 @@
+class Employee < ApplicationRecord
+  taskable_assignee
+end