+++ spec/todoapp/app/models/user.rb
@@ -1,0 +1,3 @@
+class User < ApplicationRecord
+  taskable_assignee
+end